+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 - Added `ViewAware.OnViewDetached(view, context)` and `IViewAware.GetViews()` 
-- Added `ViewModelTypeConfiguration` 
+- Added `ViewModelLocatorConfiguration` 
 - WPF: Added `PageLifecycle`
 - WPF: Added `IWindowManager.ShowOpenFolderDialog(settings, ownerViewModel)`
 - WinUI: Added `WindowLifecycle`, `PopupLifecycle` and `PageLifecycle`
@@ -19,9 +19,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Enabled nullable
 
 ### Removed
-- Removed `SimpleContainer`
+- Removed `SimpleContainer` and `ViewModelTypeResolver`
 - Removed `IViewAware.ViewAttached` and `ViewAware.Views`
-- Removed `ViewModelTypeResolver.AddMapping<TView, TViewModel>(context)`
 - WinUI: Removed `SuspensionManager`, `NavigationService` and `FrameAdapter`
 
 ### Fixed